Simulation of DNS functionality.
This project shows how DNS takes the URL and sends the request in the
form of packets( query ) to Local DNS server and then to Root DNS
server. The Root DNS redirects to the Host DNS server. Now, If the file
exists then it gets shown on our browser by following the same path.